fix(PoolClient) : Refactor DTO and Imports from pool to pool client
