+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/component/saas/dto/ImportResponsePage.kt
@@ -21,9 +21,9 @@ package org.eclipse.tractusx.bpdm.pool.component.saas.dto
 
 
 import org.eclipse.tractusx.bpdm.common.dto.response.AddressPartnerResponse
-import org.eclipse.tractusx.bpdm.pool.dto.response.AddressPartnerCreateResponse
-import org.eclipse.tractusx.bpdm.pool.dto.response.LegalEntityPartnerCreateResponse
-import org.eclipse.tractusx.bpdm.pool.dto.response.SitePartnerCreateResponse
+import org.eclipse.tractusx.bpdm.pool.api.dto.response.AddressPartnerCreateResponse
+import org.eclipse.tractusx.bpdm.pool.api.dto.response.LegalEntityPartnerCreateResponse
+import org.eclipse.tractusx.bpdm.pool.api.dto.response.SitePartnerCreateResponse
 
 data class ImportResponsePage(
     val totalElements: Int,
