+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/component/saas/service/PartnerImportService.kt
@@ -21,7 +21,7 @@ package org.eclipse.tractusx.bpdm.pool.component.saas.service
 
 import jakarta.persistence.EntityManager
 import mu.KotlinLogging
-import org.eclipse.tractusx.bpdm.pool.entity.SyncType
+import org.eclipse.tractusx.bpdm.pool.api.dto.SyncType
 import org.eclipse.tractusx.bpdm.pool.service.SyncRecordService
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Service
