+++ bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/dto/ChangelogEntryDto.kt
@@ -19,8 +19,8 @@
 
 package org.eclipse.tractusx.bpdm.pool.dto
 
+import org.eclipse.tractusx.bpdm.pool.api.dto.ChangelogType
 import org.eclipse.tractusx.bpdm.pool.entity.ChangelogSubject
-import org.eclipse.tractusx.bpdm.pool.entity.ChangelogType
 
 data class ChangelogEntryDto(
     val bpn: String,
