--- bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/dto/request/PaginationRequest.kt
@@ -1,40 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021,2023 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ******************************************************************************/
-
-package org.eclipse.tractusx.bpdm.pool.dto.request
-
-import io.swagger.v3.oas.annotations.Parameter
-import io.swagger.v3.oas.annotations.media.Schema
-import jakarta.validation.constraints.Max
-import jakarta.validation.constraints.Min
-import jakarta.validation.constraints.PositiveOrZero
-
-@Schema(name = "PaginationRequest", description = "Defines pagination information for requesting collection results")
-data class PaginationRequest (
-    @field:Parameter(
-        description = "Number of page to get results from", schema =
-        Schema(defaultValue = "0"))
-    @field:PositiveOrZero
-    val page: Int=0,
-    @field:Parameter(description = "Size of each page", schema =
-    Schema(defaultValue = "10"))
-    @field:Min(0)
-    @field:Max(100)
-    val size: Int=10
-)
