--- bpdm-pool/src/main/kotlin/org/eclipse/tractusx/bpdm/pool/dto/response/AddressMatchResponse.kt
@@ -1,31 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2021,2023 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ******************************************************************************/
-
-package org.eclipse.tractusx.bpdm.pool.dto.response
-
-import io.swagger.v3.oas.annotations.media.Schema
-import org.eclipse.tractusx.bpdm.common.dto.response.AddressPartnerSearchResponse
-
-@Schema(name = "AddressMatchResponse", description = "Match with score for a business partner record of type address")
-data class AddressMatchResponse(
-    @Schema(description = "Relative quality score of the match. The higher the better")
-    val score: Float,
-    @Schema(description = "Matched address business partner record")
-    val address: AddressPartnerSearchResponse
-)
